connect ft-quadratic tests to previously deployed contracts
